Make docs work again (#11)
Removed a bunch of unnecessary stuff, and brought in the
webpack.config.js from our proprietary repo, slightly modified. We’ll
have to do some work later to deduplicate this code, but this will do
for now.

This will set us up to easily move more file into this open source repo.

Also renamed all the doc files to use the `.mdx` extension, to
differentiate them from the `.md` extension for which we want to use
the `raw-loader` (in code that is not present in this repo yet). And
added a period between the numbers and document title, while I was at
it. :-)

Test plan: verified that this works locally.