Lines: workaround for WebGL/Linux bug (#24)
A couple users reported bugs with line rendering on certain Linux laptops. I was able to narrow it down to a problem with `stride`/`offset` when multiple vertex attributes are views on the same underlying buffer.

This pure-regl example demonstrates the problem. It renders a simple moving average of some data: https://codepen.io/anon/pen/maVppw?editors=0010

On a Mac and some other platforms, this renders as expected (green = stride of 4 bytes, red = stride of 8 bytes):
https://user-images.githubusercontent.com/14237/49979809-94225f00-ff05-11e8-8237-c8b87822b4ee.png

However on the affected hardware, the result differs:
https://user-images.githubusercontent.com/14237/49979841-b916d200-ff05-11e8-9c23-1df3dddefdff.png

This is consistent between Chrome & Firefox, so it seems like it might a graphics driver/hardware bug.

The problem seems to occur when 3 or 4 attributes are using the same buffer, but not with only 2, so we can work around it by taking a slight perf hit to upload 2 copies of the same data into two buffers so that each has only two attributes reading from it.