Worldview: fix hitmap bug (#326)
I'm seeing an error in our logs: "(regl) can not resize a framebuffer
which is currently in use" in the readHitmap function. Fix by debouncing
the readHitmap function because the framebuffer can only be used in
one function at a time. Also add caching to our readHitmap function: if
we haven't repainted or changed the pixel that we're looking at, we return
the previous result.

Test plan: test added for queuePromise for caching. Testing that caching
actually works (an implementation detail) seems a bit difficult right now
given our test harness.

Release plan: minor release required