Add stories setup config flow for webviz-core (#345)
## Summary

We'd like to be able to run custom setup code for storybook to support new features we're adding internally. For now it'll just be a no op.

## Test plan

Tested build and storybook locally, brought over internally repo and confirmed we can use this change to address our current issues.

## Versioning impact

No, only a test change.