Attempt to fix "invalid height" regl errors (#435)
This error occurs when a texture is beyond the max texture size. It
only shows up on certain linux machines right now, and started in
the past month. It's possible that this error is because of some
error in the messages we're seeing, but a more likely reason is the
launch of the gltext feature.

We are creating an atlas in GLText with a limited width, which means
that the height direction grows much larger than the width. It's
possible (not reproduced) that this causes the height to exceed
GL_MAX_TEXTURE_SIZE on some machines. Instead set the width of the
atlas to the actual GL_MAX_TEXTURE_SIZE, so that we give our
characters more horizontal space and hope that we don't exceed the
height limits.

Test plan: hope