Do not require a scene in GLTF (#524)
* Do not require a scene in GLTF

The `scene` property is optional. If it's not specified, use "the first"
scene.

In a .glb file I opened that didn't have a specified scene, there was
one key in the `scenes` object: `"0"`. In GLTF 2 `scenes` is an array, so
we might be able to get away with just `scenes[0]`, but this seems
slightly more robust.

I don't think we need to be super careful about files without scenes.
Maybe we want some controls for the scene to draw to be specified in the
command if there could be more than one, but let's not worry about that
until we come across it.

* Throw if there are no scenes to render